ready/requirements.txt
torch>=1.13.0
torchvision>=0.14.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.62.0
click>=8.0.0
Pillow>=8.3.0
onnxruntime-gpu>=1.19.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
timm>=0.6.0
psutil>=5.8.0  # System monitoring (CPU, memory, etc.)
pynvml>=11.0.0  # Optional: for GPU utilization monitoring